Changed config package internals
Minor refactoring to improve code logic and readability.

There was a try to refactor the ConfigSourceBase self
_parameters away but this failed. The class now has the
same parameters in attributes and _parameters dictionary.
The problem was that the internal logic relies on some
cases existence of some configuration parameters and this
is error prone to change. It was defined to leave the
implementation as is for now.

The try/catch block was not needed in few places that now
are sure to use str in Python 2 and Python 3 and do not
cause exceptions anymore.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>