Fixed created content not being stored
Previous refactoring broke basic functionality to store
created content into database that is located in disk. The
default was accidentally set for database to be created
into memory. Because each session terminates the database
session, the content was never persisted.

The new way of configuring the storage and database schema
location is simpler than before. There were unnecessary
complexity with parameters being stored into Config() that
were not really needed.

There refactoring is not complete because there are the
unit tests for Sqlite database that are currently still
failing and being skipped.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>