Fixed database schema file path config
The Sqlite schema is always installed by the tool. This
means that the file must be there always in specific path
and there is no need to check first if the path exist.
The schema file can be checked directly.

This is different from database storage file which is not
existing in the first run.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>