Fixed failing tests in Python 2
The Sqlite unit tests left database on a disk after each
test. This affected to sequential test. In case of Python 3,
the database is in memory which gets terminated automatically
after no connections are left. Python 2 uses disk DB for tests
since Sqlite3 in Python 2 does not support in-memory DB.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>