Changed configuration source digest
Internal refactoring and there are no external changes.

The digest is now changed to new more Pythonic way of
accessing the data. This means gradual removal of the
getters that bloat the configuration classes.

There is now inconsistency because with Digest (string),
None means that it was not given by config source. But
with content data, an empty tuple means that it was not
given by config source.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>