Fixed configuration source options
There were few bugs in the new implementation that were masked
by old implementation. These are now corrected.

The tool must be aware if specific parameter were given to
Config() source base class. These parameters are: digest,
data, sall, stag and sgrp. The default value for these must
be None or they must be left out from the parameter dictionary
give to the class inherited from the Config source base class.

There is an open fault in Pylint /1/. It does not see the usage
of property decorator with self._parameter when the __init__
has been set with self.parameter. The later is valid syntax
for the Python property decorator /2/.

/1/ https://github.com/PyCQA/pylint/issues/409
/2/ https://www.python-course.eu/python3_properties.php

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>