Changed configuration source options
Internal refactoring and there are no external changes.

Converted the operation filename and file type setting. These
methods were left to Config() because they require knowledge
of parameters like operation and content type which would
have to be passed to Parser. It was considered that the
basic filename reading from user is the configuration base
class task but more sophisticated parsing of the filename
is left to Config()

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>