Removed duplicated initializers
Removed default initialization in configuration base class
that was twice.

It was also noticed that the lstrip() in ConfigSourceBase
and its __repr__ setting was causing 2-3% performance hit.
Reason is it not known. The reference test was ~0.79 when
it was with lstrip() ~0.82. Reason is not known.

Made slight changes to code comments and logs. Renamed
and clarified variable names.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>