Changed Migrate() logger setting
The Migrate class logger is now set in same manner than with
Parser(). This avoids initializing the class once since the
logger is now automatically initialized into _logger when
the class is called for the very first time. All the following
uses of the class methods will use the same logger instance.

Cleaned the Migrate from main snip.py because it was not
needed there anymore

Note! Fixed test_ut_sqlitedb_* tests when run individually.
      These tests were failing previously when run one by
      one. The reason is the inconsistency in Config() which
      creates the class variables in __init__. These are not
      seen by mocks and for mocks the class variables must
      be at the top level of class.

      This should eventually go away when/if the Config()
      is fully refactored.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>