Added customer logger handlers for Gunicorn
The Gunicorn WSGI HTTP server logs are now formatted with
the same format as the Snippy logs. The loggers are set under
'snippy' namespaces:

  snippy.server.gunicorn.error
  snippy.server.gunicorn.access

This allow same format and same control mechanism (-vv and
--debug) for Gunicorn logs.

The Snippy/Gunicorn logs contain the module in the logs. In
case of Gunicorn logs, the module is always glogging.py. It
is not sure if this can be changed to behave so that the
module would be correct.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>