Changed log formatting
The console log format was changed to improve readability
and to include server transaction identity.

The server transaction identity is now printed for each log
and it identifies one transaction which is for example
create or search operation. The intention is to help post
processing of logs to be able to combine logs for specific
transaction.

The thread, module and file line information was removed
from default console log. The intention is to bring optional
JSON formatting that contains more information.

Enforced also all logs explicitly to be lower case.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>