Added JSON formatted log string
It is now possible to output log strings in JSON format
with option --json-logs. The intention is to use this with
server and push the logs to stdout. This is the default
method to print the logs in containers.

The JSON log container more information fields than the
text log. The purpose is that the JSON logs are used with
the server and the logs are collected in a centralized
server or database where they are post processed.

The --json-logs can be used with --debug or -vv option.
With -vv option, the message field is truncated in same
manner than with text logs. With the --debug option, the
message field is not modified.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>