Fix JSON API response with empty response
Based on JSON API v1.0 specification, if the request is made
to specific resource which is not found, the response must have
the top level data JSON object set to null /1/. This was set
incorrectly to string presentation from 'null' which is not
processed correctly by JSON library. If the JSON structure has
value null, it gets correctly transformed to None in case of
Python.

/1/ http://jsonapi.org/format/#fetching-resources

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>