Fix self link in REST API resource response
If specific resource is requested, there is the top level
JSON object links that contain self which links to the
resource. This was previously sent incorrect if the API
request updated the resource.

Now the link is returned correct and it is pointing to the
updated resource.

The top level links JSON object is not always present. It
cannot be currently set correct if the 'digest' field is not
included in the response. This is the case when user request
explicit fields and does not list the 'digest' field.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>