Change Config() initialization
Internal refactoring and there are no external changes.

Moved the Config() initialization at the very beginning of
the process startup. The Config() is initialized with the
first set of command line arguments.

This allows using the Config to provide for example log
level settings. This was not previously possible with few
parameters because the Config() was initialized in so late
phase. This removes the need to access the sys.argv directly.

The code is also now better because the mocking the the main
level run methods can be done without patching sys.argv list.
Now the parameters can be just passed to Snippy() or run_cli().
The intention is to allow both Snippy() and run_cli() to get
command line parameters if there is a need to run sequential
CLI commands. This is common pattern in test cases.

Code cleanup is still needed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>