Remove sys.argv from test cases
The sys.argv is not needed anymore because the arguments
are passed from main() as a parameter. This eases testing.

There are few exceptions that are not sorted yet. The help
command has additional parameters 'tests' and 'examples'
which requires checking of the sys.argv.

There is one test commented out. It may be that the exception
set by the side_effect for ImportError turns this setting
global. This affects to following test.

/1/ https://github.com/pytest-dev/pytest-mock/issues/84

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>