Add release of resources if argparse exits
Argparse Python module exits in case the argument itself
causes exit (like --help or --version) or parsing fails.
This is now captured and more controlled exit is made.

This is made because more controlled way to release the
resources was needed and wanted. Now the code runs always
in main() to controlled release of resources.

There was a single test case which for some reason did
not cause reset of Cause(). This leaked the cause to
next case. It was considered that it would not be good
to reset (or init) the Cause() in the start but to
guarantees that the code flows always to end to the
controlled release of resources.

The reason why the test_console_help_tests_no_package
in Console tests leaks the cause to next test is not
known. The code exist in that case and snippy.reset()
is not possible in the exception branch of this test.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>