Fix failing tests when run individually
Few tests were run OK when all the tests were run but failed
when run individually. The reason is that test helpers use
Snippy() Content() methods to get data templates. In order
to use this helper method, it requires initialized Config().

It may have been better to fix the helpers but as of now,
the Snippy() is initialized before the get_template() helper
is called.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>