Change test cases to use pytest tear down method
Tests are in most cases using pytest framework. Changed most
of the test to use teardown class from pytest instead from
unittest package.

This may have been reason why it seemed that one test failure
can cause sequential tests to fail. Each test should have tear
down method that releases the database resources in case of
failure. It seemed that this would not always work. This has
not been verified if this change corrected this. This problem
is not currently visible because the flag in pytest that stops
the test immediately in the first failure.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>