Change log severity level name format and order
The text log format was changed. The debug level name is now
printed with one letter and the operation ID is printed after
the level name.

Also severity names are now according to RFC5424.

It was considered that the one letter log level name is clear
enough. This solves the problem of static width "header"
before the actual message. Author just thinks that it is more
pleasant to read the logs with the log is constant width before
the actual message.

In case of JSON formatted logs, the level names are in full.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>