Add configurable server base path
It is now possible to configure the REST API server base
path with: --base-path /snippy/api/

The base path must start and end with a slash. Missing of
leading or trailing slashes is correct from code but other
failures will result usage of default server path.

The error checking of the base path is very simple since it
is hard to make bullet proof. At currently there is no known
way to make this rock solid. Using os.path result different
results in Linux and Windows. URL parsing checks against
HTTP URL which is not the case in here because this is Falcon
WSGI server route. Copying all the same checks than in Falcon
would be one approach but it was considered not good at this
point of time.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>