Add REST API documentation
The Swagger API is now converted to Sphinx documentation with
sphinxcontrib-openapi plugin. The autogeneration might not yet
contain all the needed information but this was best that was
found.

The reason to use Sphinx is to try to bundle the API documents
into same document where the rest of the tool is document. User
can always browse the online Swagger API to get the Swagger
GUI.

Embedding the Swagger API documents as a separate HTML generated
by one of the many conversion programs were not pleasing authors
eye. Single page simple API HTML document would have been useful
also with the plans to include the API and short document also
to the Github pages. But so far no good candidate has been found.
The Sphinx converted document is simple and good when the examples
are added into it.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>