Fix Read the Docs build
The new dependency for sphinxcontrib-openapi is missing from
Read the Docs (RTD). It seems that RTD runs the default install
which does not include the Sphinx plugin.

The new dependency was not wanted to default requirements.txt
so new one was created. It seems that RTD has configuration
setting to define this file.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>