Add simple JSON schema validation
There is now first version that validates the JSON format
according to JSON API specification v1.0. This is just a
very simple check that tries new dependency Schema.

It was too difficult to get the local file references to
work with JSON schema (python jsonschema). There is an
open fault that made getting the auto generated JSON
schemas from OPEN API just not work.

The Python jsonschema also felt a bit sluggish. But this
was not properly verified since it was not working. The
python Schema module seems pretty fast even though there
was 7-10% performance hit in reference API test. It was
not completely sure if this change or some recent change
caused this.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>