Add JSON schema validation for PUT request
The initial validation of JSON schema is now implemented for
collections (POST) and resources (PUT). Basic data structure
is checked against the JSON API specification v1.0. The top
level data object and the type checks and enforces now the
type as 'snippet' or 'solution'.

The code was refactored to hide the schema validation if
there is a better version or option from it.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>