Add explicit versions for all external packages
External Python packages all have now explicit versions. The
tool may work with different versions but as of now versions
are set exactly. If there are problems, these can be relaxed
for example so that the there is a range of versions.

For example Falcon is planning to release 2.0.0 version which
will drop Python 2 support. Because of this kind of reasons,
setting explicit dependencies is safer.

The side effect may be that already installed dependency is
with different version.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>