Add configurable database location
It is now possible to define the location of the database.
This allows for example server to run in freely defined
location.

  $ snippy import --defaults --storage-path /home/temp
  $ snippy --server --storage-path /home/temp -vv

The storage path must be configured for each start of the
tool. That is, for CLI mode, the parameter must be there
always if location of the database is changed. It is possible
to create alias for the configuration that always defines
the storage path. There is no configuration file that would
persist this configuration.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>