Fix OpenAPI definitions
Removed HTTP status codes that were not needed in the code.

Checked and fixed HTTP return codes in OpenAPI specification.

Refactored OpenAPI specifications to have common parameters
that can be used by multiple operations. This was preparation
to bring the solutions into the API definition. The intention
is to wait and see a bit so that the snippet part stabilizes
so that there is no need to fix same issues in OpenAPI
definitions in multiple places.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>