Fix Falcon WSGI server media type
Falcon server was not accepting 'application/vnd.api+json' as
content type. The default content type is 'application/json'
in Falcon. The later content type is still accepted and the
Snippy server will still respond to it. This is against JSON
API specification but as of now this is accepted.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>