Fix Logger documentation
Some parts of the documentation was decided to be moved to
auto generated rst documentation. The intention is that longer
descriptions of classes and rules are added to rst document.
The intention is not to bloat the code with too many comment
lines. This may result unsynchronized code and documentation
but the rst document is intended to be kept on high level
guidance and service description document.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>