Fix exception when importing and exporting
There was a long standing issue which may have caused
exceptions when importing and exporting. The json and
yaml modules were imported at the last minute just before
usage. But the same try/catch contained other branches
which may have failed and the exception handling tried
to catch also the yaml exception which failed.

The reasoning previously to include the import of json
and yaml in last minute was to somehow optimized the
speed. But this was premature and failed optimization.

Now the imports are at the top level.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>