Change get_utc_time() mocking to be more centralized
This refactoring tries to make the amount of calls that needs
to be mocked for get_utc_time() more centralized. This tries
to avoid having to patch all tests when the call count changes
to get_utc_time().

One of the problems here is that there are different amount oif
calls to get_utc_time() depending how the content is added in
each test case. The add_defaults() and add_one() helpers call
Snippy code in order to make things easier (laziness). This
complicates calculating the count of calls to get_utc_time()
which is different depending on how the data is added.

The new way of hacing constants for the mock should be at
least a bit better than old where all test made the same.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>