Add more complex test with pytest fixtures
There is now two tests and the default snippet content can
be added with fixtures. This removes much of the patch clutter
but it requires every test to have its own function. This is
actually good because it will help to run individual tests and
makes the setup and cleanup easier.

This should be much better than earlier implementation.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>