Fix resource release failure with pytest fixture tests
New tests with fixtures must not create the Snippet() by
themself. If test do this, the resources are no properly
released. There was accidentally a Snippy() created in
the solution test and this had side effects.

Also the new fixture tests must not call the Database
delete_storage method because otherwise the database
gets deleted before the finalizer is called which is
currently failing. The finalizer could handle this error
but it is considered now that it is easier to find faults
in tests if this is not managed in the delete_storage.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>