Change timestamp handling for content management
Clarify rules to update timestamp in documentation. Change
the updated timestamp to reflect the new rules. That is, in
case of text file import, the updated and created timestamps
will get lost because there are no timestamps in the text
format output.

Tests are failing and this is intermediate commit to continue
later.

[skip ci]

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>