Fix installer
There was a typo in setup.py that prevented installing default
templates. This caused a situation where normal 'snippy --help'
did not result any output.

The second fault was that in this case the tool made exit
without any indication to the user. The reason was that the
default is that not even error logs are printed to user. Now
this is improved so that in case there is internal error that
leads to tool exist, the exit cause is printed for user.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>