Change release tag to 0.8.d (development)
Updated TODO list with the findings and learnings from this
release.

One learning was that if the tool is installed without the
--user option in pip, all snippy commands must be run with
sudo. There is no even read access to files e.g. in the data
directly by other users. The thought was that this behaves
differently now with the first wheel distribution. But the
explanation sounds logical and perhaps this was just missed
last time. A bit more research is needed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>