Fix exit log in case of fatal failure to start
There are three cases that are considered fatal cases which
prevent the tool to run. These are missing or inaccessible
data files that must be installed with the tool.

Now all three places produce proper error log before exit.

If the tool is already about to exit with a help text from
the CLI parser, the log is not printed. This additional
error log is prevented because it was considered confusing.
In this case the exit is already being made due to CLI parse
failures like option errors or due to normal reason like with
the --help option.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>