Add method for log setting
The plan was to update the log settings in Config.load() but
it was realized that the initialization of Logger does not
currently do anything any changes to existing loggers after
the first initialization.

The logging and development settings were moved to method
to be later updated in load() when the Logger is updated
to update existing loggers with new settings.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>