Change console tests to use Pytest fixtures
Few tests were moved to use new Pytest fixtures. How ever,
there may be hidden problems. The change required parameter
in the snippy fixture.

It is not understood why 'test_very_verbose_option_001' has
few log lines in stdout.

It is not understood why 'test_debug_option_001' has exceptions
in stderr. The exceptions are on the Logger which complains
'I/O operation on closed file'. This seems to work and it
seems to be somehow incorrect usage of Pytest fixtures, caplog
and capstd. But as of now, this is a mystery.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>