Change CLI option tests to use Pytest fixtures
There are few TODO cases left for the CLI option tests.
There are two strange cases documented with TODO that
produce errors in the tests. In reality, the same errors
seem not to exist. Current assumption is that there is
some hidden conflict in new pytest fixture mocks that
cause these issues in CLI option tests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>