Fix importing content for tests with Python 3.4
The Python 3.4 branch was failing in Travis. The reasons is
unknown. This seems to be related to mocking file open in
Migrate and to the 'create=true' value that was removed.
For some reason it seems that this is needed in Python 3.4.
The parameter just defines that the mocked attribute is
created if it is not existing.

The reason why this is failing two times in a row only in
Python 3.4 is not known.

  mocker.patch('snippy.migrate.migrate.open', mocked_open, create=True)
  cause = snippy.run_cli(['snippy', 'import', '-f', 'content.txt'])

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>