Fix false positive in test case
The test was not testing what it should have. The reason was
that some cases like the --help examples still uses sys.args
instead of command line parameters as a method parameters.
And the setting of sys.args was missing in this case. The
reason is for this misuse is missing refactoring.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>