Fix unnecessary help text in test cases
When test cases were run previously, there was always a help
text printed. This was because the Snippy() object was created
with only one parameter which was the tool name. This must
print normally the help text.

The fix was to set a quiet parameter in tests if there are
no other parameters. The quiet parameter is now also dynamic
because CLI tests are normally testing the tool output. If
the quiet would be set only at the startup, the tested output
would be blocked.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>