Fix tests to use snippy.run()
Test now use generic run() method that hides whether a CLI
or server session is run. The run_cli() and run_server()
are now internal methods and not to be used externally.

Cleared also the Pytest fixture usage. The snippy fixture
was unnecessary defined in test case fixture when it is
just enough to define it as a test case parameter.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>