Fix setup to use UTF-8 when reading README.rst
The setup.py now explicitly uses UTF-8 when the README.rst
is read. This tries to avoid problems in installation. There
are no currently non ASCII characters in the README file even
when the system default encoding would fall into ASCII.

This should help also in case other encodings are used. But
this is not completely understood if this really is the case.
That is if non ASCII or UTF-8 encoding is used by default,
installer might try to read the ASCII characters with other
encoding which would fail.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>