Add tests for JSON API v1.0 media type in request
All the API tests previously used only 'application/json'
in the request that is not explicitly accepted by the JSON
API v1.0. Added tests that verify also:

  1. application/vnd.api+json
  2. application/vnd.api+json; charset=UTF-8

in accept field of the request header. The later alternative
with charset is also not allowed by JSON API v1.0. This is
how ever an exception to JSON API v1.0 rules that the server
allows these two options on top of 'application/json'.

The reason to allow these options currently is that author
does not understand what would be reasons to reject these.
The 'application/json' might be commonly used by the users
so it tries to make using of the server easier. Also the
charset parameter is accepted to make it self evident that
the returned data is always UTF-8 encoded.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>