Fix tests if gunicorn is not installed
Test were failing if gunicorn was not installed. The same is
not yet applied to Falcon which would need the same fix.

There are something strange with Falcon. When Falcon is
uninstalled with pip, the import still works and the pytest
importorskip seems not to work. This might be environment
specific problem where something else is creating the falcon
namespace for Python.

It would be also good if importskip from Pytest would support
list of modules. Now the first parameter must be a string
and it seems that list of modules is not supported.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>