Change --profile option internals
The profiler now uses the --profile option read from command
line from argparse and not the sys.argv directly. This tries
to make code more logical and to avoid shortcuts that may cause
surprises when code is developed.

This also should be better with possible conflicts from other
arguments and it allows better error handling as well.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>