Change --help option internals
Customised help actions previously used sys.argv directly.
This was considered as a bad design which may have led to
problems. Because of this, the code logic now avoids using
sys.argv.

Side effect is that the --no-ansi option now works only
when set before the --help option. The reason is it that
the argparse module stops parsing arguments immediately
when the --help option is met. There was no solution found
to overcome this without using sys.argv.

It was considered that the arguments must be always used
through argparse and this side effect is accepted. This
way allows better control and sanity checks for the command
line options.

No external changes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>