Change Logger() configuration internals
Removed the usage of sys.argv from the Logger(). Now only the
main() provides the sys.argv for the tool. This tries to prevent
misuse and parse failures of command line arguments by enforcing
usage of parameters through Config() that uses the argparse
module.

Only the Logger initialization reads directly from args received
from the main level. The reason is that the Logger configuration
must be set as early as possible in order to produce logs from
command line argument parsing. But still, this does not use the
sys.argv directly.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>