Fix server log configuration
Previously the server log configuration was kept only for
the first operation. This bug was made when all log settings
were made dynamic. The problem is that the configuration is
read for all the API operation which includes also the log
configuration. This resets the log config to default after
the first operation.

Now only the quiet flag is updated dynamically. The reason
for this is that the test cases depend on this functionality
to prevent unnecessary help text when the Snippy() object
is created by Pytest fixtures.

The behaviour is:

  1. CLI runs once and the -q is applied to one run.
  2. Server startup logs can be prevented with -q flag.
  3. Server admin API will be implemented to change settings.
  4. Snippy() as a library requires -q for every run() method.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>