Remove unnecessary content template check
The content test against default template was made in two
places when it is needed only once. The template test for
the content is now made just before insert.

This also reduced the amount of calls to system time.

Changed also the cause codes related to template and data
checks to better reflect the failures.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>