Fix unnecessary call of system time
The Content.convert_text() is called with new content or
already created content is exported. When new content is
created, the created and updated timestamps are always set.
This guarantees that the created time can be always used
directly from Content() and there is no need to call system
time.

When the template DATE tag is updated, it always means
that new content or template is being created. In this
case the created timestamp contains always correct time.

Also if the system time would be read again to replace the
DATE tag, it would create slightly different timestamps in
the metadata and in the text template. Now the timestamps
are same in metadata and in text template.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>