Change created Content() to update digest
The digest is now updated automatically by the Content()
object when the content is created. This is part of changes
that are planned to simply inserting data into the database.

Also this feels a better way to encapsulate this operation
into the Content() object and remove the need from user
to remember to update the digest according to created or
updated content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>