Change JSON timestamp accuracy
The JSON ISO8601 formatted timestamp now uses microsecond
granularity even when the current implementation supports
only millisecond granularity. This is made to avoid changes
in end user parser when the microsecond granularity is
implementation. This is likely not investigated or added
when Python 2.7 is supported.

Fixed also Python 2.7 test case that verified then ISO8601
timestamp format. The %z directive is available only from
Python 3.2 and onwards.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>