Fix JSON log timestamp microseconds
The ISO8601 formatted timestamp in JSON logs had incorrect
microseconds. Previous change used the millisecond portion
incorrect in least significant numbers when the msecs part
should be most significant digits in microsecond.

It seems that the logging Record and the msecs field contain
milliseconds in floating point number. It seems that digits
after point represent more accurate timestamp than mseconds.
Because of this, the timestamp should be now in microseconds.

Removed also the import exception check for OrderedDict since
this module is by default in Python 2.7 which is the oldest
supported Python version.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>