Add command line option --log-msg-max
It is now possible to define the log message length with
new --log-msg-max option. The option accepts integer value
and uses default length if parsing fails.

This is not currently dynamically updated value after
configuration. The plan is to have new API for server to
update server configuration which allows updating for
example the logger.

The Logger() itself supports dynamic configuration updates.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>