Fix configuration source parsing
The configuration source was read only in case CLI was run.
This was problematic with some tests where the server was
started first, and the received REST API query like DELETE
/snippets did not result reading the configuration again.

The problem was that if the server mode is not updated, the
sequential runs that need to run the CLI command, like for
exporting the data for testing purposes, will fail because
the code runs the server.

Now the configuration is read always again in run() if the
arguments are provided. This is also more logical like this
and it forces the update of arguments also when running
server.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>