Remove DELETE on collections
It is not possible to delete content based on query parameters
in DELETE request. Previously the query must have matched to
only one content to trigger the delete. But this was considered
confusing.

It is not fully understood, but there might be problems with
caching if the delete operation is based on query parameters
and the delete resource is not visible in request URL like is
the normal case. This may cause difficulties in intermediate
caches which may not be able to see that some resource have
been delete.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>