Add support for PATCH request
PUT and PATCH request behave now as they should with fields
that can be modified. The PUT updates all the fields and
PATCH can be used to update only selected fields. This is
the normal behaviour of these requests.

If PUT request does not contain some fields, a default
value is applied. For example the default value for the
'brief' field is empty string.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>