Change Content() internals
Cleaned the content.py code layout. Override Content() equal
and not equal comparisons to allow comparing the content.

Moved time stamp calculation outside of the Content(). The
goal is to be able to remove Config() dependency from Content()
in order to allow importing Content() without circular
dependencies. Now the Content() is created on Snippet() or
Solution() because the Content() cannot be created in Config()
due to circular dependencies.

The Content() is still bit awkward but this is hard to clean.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>