Add documentation how content attributes are updated
The content attributes and their behaviour is now tested
and documented. There are some attributes which are freely
modified and some attributes that are automatically defined
by the tool.

The digest, updated, created and category attributes cannot
be modified by user. These attributes remain in their initial
values or they are automatically updated.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>