Fix validation of HTTP request with empty content
Sending empty data like: {'data': []} was accepted as valid
request. This should not be accepted and is now rejected with
400 bad request status.

Changed also JSON schema validation from Python schema to
Python jsonschema package. The later seems to be more active
and it was able to handle this failure case. Also the syntax
for the later seem to be more readable and follows standards.

How ever, there seems to be roughly 5-10% performance hit with
the new jsonschema package. This is considered as an OK trade-
off to gain more robust validation of received JSON data.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>