Add new way to parse sorted fields from REST API
This is a preparation work to allow REST API pagination. The
sort based on specific attributes is possible directly from
Sqlite DB. This allows easy way of requesting one page based
on offset and limit from the database.

The intended implementation does not handle case where new
content is added between requesting multiple pages. But it
is still considered straightforward and simple which is very
good solution for very small scale usage.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>