Change content sorting internals
Content sorting based on sort parameter in URL was changed
to be done by the Sqlite database. This is easier and likely
more robust than own implementation.

This is related to ongoing effort to have content pagination
based on offset parameter from URL which can be also done
with the help of Sqlite database.

No externally visible changes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>